# openapi.yaml
openapi: 3.1.0
info:
  title: Trustify
  description: Software Supply-Chain Security API
  license:
    name: Apache License, Version 2.0
    identifier: Apache-2.0
  version: 0.4.0-beta.1
paths:
...
  /api/v2/group/sbom:
    get:
      tags:
      - sbom_group
      summary: List SBOM groups
      operationId: listSbomGroups
      parameters:
      - name: q
        in: query
        description: |
          EBNF grammar for the _q_ parameter:
          ```text
              q = ( values | filter ) { '&' q }
              values = value { '|', values }
              filter = field, operator, values
              operator = "=" | "!=" | "~" | "!~" | ">=" | ">" | "<=" | "<"
              value = (* any text but escape special characters with '\' *)
              field = (* must match an entity attribute name *)
          ```
          Any values in a _q_ will result in a case-insensitive "full
          text search", effectively producing an OR clause of LIKE
          clauses for every string-ish field in the resource being
          queried.

          Examples:
          - `foo` - any field containing 'foo'
          - `foo|bar` - any field containing either 'foo' OR 'bar'
          - `foo&bar` - some field contains 'foo' AND some field contains 'bar'

          A _filter_ may also be used to constrain the results. The
          filter's field name must correspond to one of the resource's
          attributes. If it doesn't, an error will be returned
          containing a list of the valid fields for that resource.

          The value 'null' is treated specially for [Not]Equal filters:
          it returns resources on which the field isn't set. Use the
          LIKE operator, `~`, to match a literal "null" string. Omit the
          value to match an empty string.

          Examples:
          - `name=foo` - entity's _name_ matches 'foo' exactly
          - `name~foo` - entity's _name_ contains 'foo', case-insensitive
          - `name~foo|bar` - entity's _name_ contains either 'foo' OR 'bar', case-insensitive
          - `name=` - entity's _name_ is the empty string, ''
          - `name=null` - entity's _name_ isn't set
          - `published>3 days ago` - date values can be "human time"

          Multiple full text searches and/or filters should be
          '&'-delimited -- they are logically AND'd together.

          - `red hat|fedora&labels:type=cve|osv&published>last wednesday 17:00`

          Fields corresponding to JSON objects in the database may use a
          ':' to delimit the column name and the object key,
          e.g. `purl:qualifiers:type=pom`

          Any operator or special character, e.g. '|', '&', within a
          value should be escaped by prefixing it with a backslash.
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: |
          EBNF grammar for the _sort_ parameter:
          ```text
              sort = field [ ':', order ] { ',' sort }
              order = ( "asc" | "desc" )
              field = (* must match the name of entity's attributes *)
          ```
          The optional _order_ should be one of "asc" or "desc". If
          omitted, the order defaults to "asc".

          Each _field_ name must correspond to one of the columns of the
          table holding the entities being queried. Those corresponding
          to JSON objects in the database may use a ':' to delimit the
          column name and the object key,
          e.g. `purl:qualifiers:type:desc`
        required: false
        schema:
          type: string
      - name: offset
        in: query
        description: |-
          The first item to return, skipping all that come before it.

          NOTE: The order of items is defined by the API being called.
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: limit
        in: query
        description: |-
          The maximum number of entries to return.

          Zero means: no limit
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: totals
        in: query
        required: false
        schema:
          type: boolean
      - name: parents
        in: query
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Matching SBOM groups
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                - total
                properties:
                  items:
                    type: array
                    items:
                      allOf:
                      - $ref: '#/components/schemas/SbomGroup'
                      - type: object
                        properties:
                          children:
                            type:
                            - array
                            - 'null'
                            items:
                              type: string
                              format: uuid
                          parent_path:
                            type:
                            - array
                            - 'null'
                            items:
                              type: string
                              format: uuid
                          sbom_count:
                            type:
                            - integer
                            - 'null'
                            format: int64
                            minimum: 0
                  total:
                    type: integer
                    format: int64
                    minimum: 0
    post:
      tags:
      - sbom_group
      summary: Create a new SBOM group
      operationId: createSbomGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SbomGroupRequest'
        required: true
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SbomGroup'
        '400':
          description: Invalid request
        '404':
          description: Parent group not found
  /api/v2/group/sbom-assignment/{id}:
    get:
      tags:
      - sbom_group
      summary: Get all group assignments for an SBOM
      operationId: getSbomGroupAssignments
      parameters:
      - name: id
        in: path
        description: ID of the SBOM
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: List of group IDs assigned to the SBOM
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uuid
        '404':
          description: SBOM not found
    put:
      tags:
      - sbom_group
      summary: Set (replace) all group assignments for an SBOM
      operationId: setSbomGroupAssignments
      parameters:
      - name: id
        in: path
        description: ID of the SBOM
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uuid
        required: true
      responses:
        '204':
          description: Assignments updated successfully
        '404':
          description: SBOM or one or more groups not found
  /api/v2/group/sbom-by-path/{path}:
    get:
      tags:
      - sbom_group
      summary: Get SBOM group by path
      operationId: getSbomGroupByPath
      parameters:
      - name: path
        in: path
        description: Hierarchical path (e.g., A/B\/C for group 'B/C' under 'A')
        required: true
        schema:
          type: string
      - name: children
        in: query
        required: false
        schema:
          type: boolean
      - name: totals
        in: query
        required: false
        schema:
          type: boolean
      - name: parents
        in: query
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Matching group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SbomGroupDetails'
        '400':
          description: Invalid path format
        '404':
          description: The group could not be found
  /api/v2/group/sbom/{id}:
    get:
      tags:
      - sbom_group
      summary: Get SBOM group by ID
      operationId: getSbomGroup
      parameters:
      - name: id
        in: path
        description: Opaque ID of the group
        required: true
        schema:
          type: string
          format: uuid
      - name: children
        in: query
        required: false
        schema:
          type: boolean
      - name: totals
        in: query
        required: false
        schema:
          type: boolean
      - name: parents
        in: query
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Matching group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SbomGroupDetails'
        '404':
          description: The group could not be found
    put:
      tags:
      - sbom_group
      summary: Update an SBOM group
      operationId: updateSbomGroup
      parameters:
      - name: id
        in: path
        description: Opaque ID of the group
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SbomGroupRequest'
        required: true
      responses:
        '204':
          description: Group updated
        '400':
          description: Invalid request or revision mismatch
        '404':
          description: Group not found
        '412':
          description: Precondition failed (If-Match header required or doesn't match)
    delete:
      tags:
      - sbom_group
      summary: Delete an SBOM group
      operationId: deleteSbomGroup
      parameters:
      - name: id
        in: path
        description: Opaque ID of the group
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Group deleted
        '400':
          description: Group has children
        '404':
          description: Group not found
        '409':
          description: Cannot delete group with children
        '412':
          description: Precondition failed (If-Match header required or doesn't match)
...
components:
  schemas:
...
    Labels:
      type: object
      additionalProperties:
        type: string
...
    SbomGroup:
      type: object
      required:
      - id
      - name
      - revision
      properties:
        id:
          type: string
          format: uuid
        labels:
          $ref: '#/components/schemas/Labels'
        name:
          type: string
        parent:
          type:
          - string
          - 'null'
          format: uuid
        revision:
          type: integer
          format: int32
    SbomGroupDetails:
      allOf:
      - $ref: '#/components/schemas/SbomGroup'
      - type: object
        properties:
          children:
            type:
            - array
            - 'null'
            items:
              type: string
              format: uuid
          parent_path:
            type:
            - array
            - 'null'
            items:
              type: string
              format: uuid
          sbom_count:
            type:
            - integer
            - 'null'
            format: int64
            minimum: 0
    SbomGroupRequest:
      type: object
      required:
      - name
      properties:
        labels:
          $ref: '#/components/schemas/Labels'
        name:
          type: string
        parent:
          type:
          - string
          - 'null'
          format: uuid
...